fix graph plotting and red teaming
