+++ evaluation/config.json
@@ -3,7 +3,7 @@
     "results_dir": "results",
     "passing_rate": 3,
     "requested_metrics": ["gpt_groundedness", "gpt_relevance", "gpt_coherence", "gpt_similarity", "gpt_fluency", "answer_length", "latency", "f1_score"],
-    "compared_models":["gpt-35-turbo", "mistralai/Mistral-7B-Instruct-v0.3", "microsoft/Phi-3-mini-4k-instruct"],
+    "compared_models":["GPT 3.5 Turbo", "Mistral AI 7B", "Phi 3 Mini 4K"],
     "max_workers": 1,
     "target_parameters": {
         "overrides": {
